fix(output): write the free-surface location flag four bytes wide again
The free-surface output wrote locationFlag as a single byte, and the Xdmf
writer declared every integer as NumberType="Int" with the precision of its
type. XDMF reads an Int as at least four bytes wide, so with the default posix
backend ParaView dropped the attribute, and seissolxdmf, which knows no integer
narrower than four bytes, returned an eighth of the values as garbage. The mesh
comparison then zipped the cells with those flags and silently compared an
eighth of them.

The flag is written as a four-byte unsigned integer again, as master did. The
Xdmf writer declares a one-byte integer as Char or UChar, which XDMF reads at
its width. The mesh comparison stops when the flags do not come back with one
value per cell.

AI-generated. Model: Opus 5

+++ src/Initializer/InitProcedure/InitIO.cpp
@@ -691,12 +691,14 @@ void setupOutput(seissol::SeisSol& seissolInstance) {
           target[0] = rank;
         });
 
-    writer.addCellData<std::uint8_t>(
+    // four bytes wide, as it has always been written: readers such as seissolxdmf know no narrower
+    // integers
+    writer.addCellData<std::uint32_t>(
         "locationFlag",
         {},
         true,
-        [=,
-         &freeSurfaceIntegrator](std::uint8_t* target, std::size_t index, std::size_t /*subcell*/) {
+        [=, &freeSurfaceIntegrator](
+            std::uint32_t* target, std::size_t index, std::size_t /*subcell*/) {
           target[0] = surfaceLocationFlag[freeSurfaceIntegrator.backmap[index]];
         });
